moved course records to csv

+++ src/main/java/golf/tracker/CourseTracker.java
@@ -3,37 +3,37 @@ package golf.tracker;
 import java.util.List;
 
 public record CourseTracker(
-        java.util.Map<Region, Integer> courseCountByRegion,
+        java.util.Map<Region, Long> courseCountByRegion,
         List<CourseRecord> courses
 ) {
 
-    public int totalCourseCount() {
+    public long totalCourseCount() {
         return courseCountByRegion.values().stream()
-                .reduce(0, Integer::sum);
+                .reduce(0L, Long::sum);
     }
 
-    public int totalCoursesPlayed() {
+    public long totalCoursesPlayed() {
         return courses().size();
     }
 
-    public int totalCoursesToBePlayed() {
+    public long totalCoursesToBePlayed() {
         return totalCourseCount() - totalCoursesPlayed();
     }
 
-    public int totalCourseCount(Region region) {
+    public long totalCourseCount(Region region) {
         return courseCountByRegion.get(region);
     }
 
-    public int totalCoursesPlayed(Region region) {
-        return (int) courses().stream().filter(cr -> cr.course().region().equals(region)).count();
+    public long totalCoursesPlayed(Region region) {
+        return courses().stream().filter(cr -> cr.course().region().equals(region)).count();
     }
 
-    public int totalCoursesToBePlayed(Region region) {
+    public long totalCoursesToBePlayed(Region region) {
         return totalCourseCount(region) - totalCoursesPlayed(region);
     }
 
     public int overallProgress() {
-        return totalCoursesPlayed() * 100 / totalCourseCount();
+        return (int) (totalCoursesPlayed() * 100 / totalCourseCount());
     }
 
 }
